requirements.txt
pypdf==4.3.1
nltk==3.9.1
moviepy==1.0.3
numpy==1.26.4
Pillow==10.4.0
